+++ semgrep-core/tests/ts/parsing/is.ts
@@ -1,0 +1,6 @@
+const isNonIdealStateProps = (object: any): object is NonIdealStateProps => {
+  if (object == null) {
+    return false;
+  }
+  return object.icon != null && object.title != null;
+};