+++ CHANGELOG.md
@@ -10,7 +10,9 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Fixed
 
-- `semgrep ci`: CI runs were failing to checkout the PR head in GitHub Actions, which is corrected here.
+- `semgrep ci`: CI runs were failing to checkout the PR head in GitHub Actions, which is
+  corrected here.
+- TS: fixed the parsing of type predicates and typeof queries
 
 ## [0.100.0](https://github.com/returntocorp/semgrep/releases/tag/v0.100.0) - 2022-06-22